refactor rows in Files into its own component
