--- shared/fs/row/containers.js
@@ -1,7 +1,0 @@
-// @flow
-export {default as Placeholder} from './placeholder'
-export {default as TlfType} from './tlf-type-container'
-export {default as Tlf} from './tlf-container'
-export {default as Still} from './still-container'
-export {default as Editing} from './editing-container'
-export {default as Uploading} from './uploading-container'
