+++ shared/fs/row/index.stories.js
@@ -5,7 +5,7 @@ import * as Sb from '../../stories/storybook'
 import * as Styles from '../../styles'
 import * as Types from '../../constants/types/fs'
 import {Box} from '../../common-adapters'
-import {WrapRow} from '../'
+import {WrapRow} from './rows'
 import ConnectedStillRow from './still-container'
 import TlfTypeRow from './tlf-type'
 import TlfRow from './tlf'
